requirements.txt
asgiref==3.3.1
Django==2.1.7
django-crispy-forms==1.11.1
Pillow==8.1.2
pytz==2021.1
sqlparse==0.4.1
tinytag==1.5.0
